+++ CHANGELOG.md
@@ -30,6 +30,14 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 - Fixed `TVGNode` that `get_max_subgraph_id` always returns the last subgraph ID. #802
 
+- Fixed `splitFasta` that source and source group order gets overriden by GVF order. #805
+
+- Fixed `summarizeFasta` and `splitFasta` being too slow. #795
+
+## Added
+
+- Added support for `--group-source` for `summarizeFasta`. #798
+
 ## [1.2.0] - 2023-08-03
 
 ### Fixed